Small tweaks

+++ pipsi.sh
@@ -3,8 +3,7 @@
 set -e
 
 package() {
-  [ -e "$HOME/.local/venvs/$1" ] && return 0
-  pipsi install $1
+  [ -e "$HOME/.local/venvs/$1" ] || pipsi install $1
 }
 
 package coursera-dl
